+++ src/test/groovy/org/codenarc/rule/junit/JUnitPublicNonTestMethodRuleTest.groovy
@@ -120,9 +120,15 @@ class JUnitPublicNonTestMethodRuleTest extends AbstractRuleTestCase<JUnitPublicN
     @Test
     void testApplyTo_TestAnnotation() {
         final SOURCE = '''
-            class MyTest extends GroovyTestCase {
+            class MyTest {
                 @Test(expected = MyException.class)
                 void shouldSendEmail() { assert count == 0 }
+                
+                @org.junit.Test
+                void otherTest1() { assert count == 0 }
+                
+                @org.junit.jupiter.api.Test
+                void otherTest2() { assert count == 0 }
             }
         '''
         assertNoViolations(SOURCE)